+++ cmd/insprctl/cli/apply_channel.go
@@ -6,9 +6,11 @@ import (
 
 	"github.com/inspr/inspr/pkg/cmd"
 	cliutils "github.com/inspr/inspr/pkg/cmd/utils"
+	"github.com/inspr/inspr/pkg/ierrors"
+	"github.com/inspr/inspr/pkg/meta"
 	metautils "github.com/inspr/inspr/pkg/meta/utils"
 	"github.com/inspr/inspr/pkg/meta/utils/diff"
-	utils "github.com/inspr/inspr/pkg/meta/utils/parser"
+	"gopkg.in/yaml.v2"
 )
 
 // NewApplyChannel receives a controller ChannelInterface and calls it's methods
@@ -16,10 +18,16 @@ import (
 func NewApplyChannel() RunMethod {
 	return func(data []byte, out io.Writer) error {
 		c := cliutils.GetCliClient().Channels()
+		var channel meta.Channel = meta.Channel{
+			Meta: meta.Metadata{Annotations: make(map[string]string)},
+		}
+
 		// unmarshal into a channel
-		channel, err := utils.YamlToChannel(data)
-		if err != nil {
+		if err := yaml.Unmarshal(data, &channel); err != nil {
 			return err
+		}
+		if channel.Meta.Name == "" {
+			return ierrors.NewError().Message("channel without name").Build()
 		}
 
 		flagDryRun := cmd.InsprOptions.DryRun