+++ pkg/api/handlers/brokers_handler.go
@@ -1,11 +1,14 @@
 package handler
 
 import (
+	"encoding/json"
 	"net/http"
 
+	"github.com/inspr/inspr/cmd/sidecars"
 	"github.com/inspr/inspr/pkg/api/models"
 	"github.com/inspr/inspr/pkg/rest"
 	"go.uber.org/zap"
+	"gopkg.in/yaml.v2"
 )
 
 // BrokerHandler - contains handlers that uses the BrokerManager interface methods
@@ -29,12 +32,14 @@ func (bh *BrokerHandler) HandleGet() rest.Handler {
 			logger.Error("unable to obtain currently available brokers on cluster",
 				zap.Any("error", err))
 			rest.ERROR(w, err)
+			return
 		}
 		def, err := bh.Brokers.GetDefault()
 		if err != nil {
 			logger.Error("unable to obtain currently default brokers on cluster",
 				zap.Any("error", err))
 			rest.ERROR(w, err)
+			return
 		}
 		brokers := &models.BrokersDI{
 			Installed: available,
@@ -43,6 +48,38 @@ func (bh *BrokerHandler) HandleGet() rest.Handler {
 		logger.Debug("current brokers:", zap.Any("brokers", brokers.Default))
 
 		rest.JSON(w, http.StatusOK, brokers)
+	}
+	return rest.Handler(handler)
+}
+
+// KafkaCreateHandler is the function that processes requests at the /brokers/kafka endpoint
+func (bh *BrokerHandler) KafkaCreateHandler() rest.Handler {
+	logger.Info("handling the brokers' kafka route request")
+	handler := func(w http.ResponseWriter, r *http.Request) {
+		// decode into the bytes of yaml file
+		var content models.BrokerConfigDI
+		err := json.NewDecoder(r.Body).Decode(&content)
+		if err != nil {
+			rest.ERROR(w, err)
+			return
+		}
+
+		var kafkaConfig sidecars.KafkaConfig
+		// parsing the bytes into a Kafka config structure
+		err = yaml.Unmarshal(content.FileContents, &kafkaConfig)
+		if err != nil {
+			rest.ERROR(w, err)
+			return
+		}
+
+		if err = bh.Brokers.Create(
+			&kafkaConfig,
+		); err != nil {
+			rest.ERROR(w, err)
+			return
+		}
+
+		rest.JSON(w, http.StatusOK, nil)
 	}
 	return rest.Handler(handler)
 }