+++ pkg/controller/client/brokers.go
@@ -14,13 +14,28 @@ type BrokersClient struct {
 }
 
 // Get gets a brokers from the Insprd
-func (cc *BrokersClient) Get(ctx context.Context) (*models.BrokersDI, error) {
+func (bc *BrokersClient) Get(ctx context.Context) (*models.BrokersDI, error) {
 	resp := &models.BrokersDI{}
 
-	err := cc.reqClient.Send(ctx, "/brokers", http.MethodGet, nil, resp)
+	err := bc.reqClient.Send(ctx, "/brokers", http.MethodGet, nil, resp)
 	if err != nil {
 		return nil, err
 	}
 
 	return resp, nil
+}
+
+// Create creates a broker into the cluster via insprd
+func (bc *BrokersClient) Create(ctx context.Context, brokerName string, config []byte) error {
+	dataBody := models.BrokerConfigDI{
+		BrokerName:   brokerName,
+		FileContents: config,
+	}
+	err := bc.reqClient.Send(
+		ctx,
+		"/brokers/"+brokerName,
+		http.MethodPost,
+		dataBody,
+		nil)
+	return err
 }